+++ hutool-core/src/main/java/cn/hutool/core/io/FileTypeUtil.java
@@ -26,10 +26,10 @@ public final class FileTypeUtil {
 	static {
 		fileTypeMap = new ConcurrentHashMap<>();
 
-//		fileTypeMap.put("ffd8ffe000104a464946", "jpg"); // JPEG (jpg)
-		fileTypeMap.put("ffd8ffe", "jpg"); // JPEG (jpg)
-		fileTypeMap.put("89504e470d0a1a0a", "png"); // PNG (png)
-		fileTypeMap.put("47494638396126026f01", "gif"); // GIF (gif)
+		fileTypeMap.put("ffd8ff", "jpg"); // JPEG (jpg)
+		fileTypeMap.put("89504e47", "png"); // PNG (png)
+		fileTypeMap.put("4749463837", "gif"); // GIF (gif)
+		fileTypeMap.put("4749463839", "gif"); // GIF (gif)
 		fileTypeMap.put("49492a00227105008037", "tif"); // TIFF (tif)
 		fileTypeMap.put("424d228c010000000000", "bmp"); // 16色位图(bmp)
 		fileTypeMap.put("424d8240090000000000", "bmp"); // 24位位图(bmp)