+++ hutool-socket/src/main/java/cn/hutool/socket/aio/MsgDecoder.java
@@ -1,0 +1,7 @@
+package cn.hutool.socket.aio;
+
+import java.nio.ByteBuffer;
+
+public interface MsgDecoder<T>{
+	T decode(ByteBuffer buffer);
+}